requirements.txt
spacy 
scikit-learn 
PyPDF2 
python-docx 
transformers 
torch